refactor(cli): extract analyzer temporal segmentation helpers
